fix(bootstrap): update import paths after link module moved to postgres/link
