+++ packages/cli/src/bootstrap/Bootstrap.ts
@@ -14,8 +14,8 @@ import ora from 'ora'
 
 import { Generate } from '../Generate'
 import { Init } from '../Init'
-import { Link, type LinkResult } from '../link/Link'
-import { LinkApiError, sanitizeErrorMessage } from '../link/management-api'
+import { Link, type LinkResult } from '../postgres/link/Link'
+import { LinkApiError, sanitizeErrorMessage } from '../postgres/link/management-api'
 import { type BootstrapStepStatus, formatBootstrapOutput } from './completion-output'
 import { detectProjectState, getModelNames, getSeedCommand } from './project-state'
 import { emitFlowCompleted, emitFlowStarted, emitStepCompleted, emitStepFailed, emitStepSkipped } from './telemetry'
